+++ utils/hct/hctdb.py
@@ -6933,6 +6933,11 @@ class db_dxil(object):
             "DXIL Container Content is well-formed",
             "In '%0', '%1' is not well-formed",
         )
+        self.add_valrule_msg(
+            "Container.UnusedItemInTable",
+            "Items in Table must be used",
+            "In '%0', '%1' is not used",
+        )
         self.add_valrule("Meta.Required", "Required metadata missing.")
         self.add_valrule_msg(
             "Meta.ComputeWithNode",