+++ docs/ReleaseNotes.md
@@ -23,6 +23,7 @@ Place release notes for the upcoming release below this line and remove this lin
 
 - The incomplete WaveMatrix implementation has been removed.
 - DXIL Validator Hash is open sourced.
+- DXIL container validation for PSV0 part allows any content ordering inside string and semantic index tables.
 
 ### Version 1.8.2407